address CodeRabbit review: memoize WSL probe, fix tautological test, extract mock helpers, surface invalid mode in JSON
