+++ test/rollout-parser.test.js
@@ -56,27 +56,9 @@ const {
 } = require("../src/lib/rollout");
 const { purgeProjectUsage } = require("../src/lib/project-usage-purge");
 
+const { mockPlatform, mockMethod } = require("./helpers/mock");
+
 const antigravityTestTokens = (text) => estimateAntigravityTokens(text || "");
-
-function mockPlatform(t, value) {
-  if (t.mock && typeof t.mock.property === "function") {
-    t.mock.property(process, "platform", value);
-    return;
-  }
-  const original = Object.getOwnPropertyDescriptor(process, "platform");
-  Object.defineProperty(process, "platform", { value, configurable: true });
-  t.after(() => Object.defineProperty(process, "platform", original));
-}
-
-function mockMethod(t, obj, name, impl) {
-  if (t.mock && typeof t.mock.method === "function") {
-    t.mock.method(obj, name, impl);
-    return;
-  }
-  const original = obj[name];
-  obj[name] = impl;
-  t.after(() => { obj[name] = original; });
-}
 
 function mockWsl(t, { distro = "Ubuntu", user = "dev" } = {}) {
   mockMethod(t, cp, "execFileSync", (_cmd, args) => {
